refactor: move app providers to a dedicated mixin
